Create Bubble Sort
In this, we start from the begnining of the array, swap the first two elements, if the first one is bigger than the second one. This process is continuously repeted until we reach the end of the array. In this way, the array elements are put in a sorted order as smaller elements bubble upto the beginning of the array.

Time Complexity : O(n²)
Space Complexity : O(1)
Sorting In-Place : Yes
Stable : Yes