Create Selection Sort
It is the most simplest algorithm, but it is inefficient. In this we linearly sacen the whole array, find the smallest element & swap it with the first element. Then again linear scan of array, find second smallest and swap it with second element. This process is continuously repeated until the whole array is sorted.

Time Complexity : O(n²)
Space Complexity : O(1)
Sorting In-Place : Yes
Stable : No